revert modularization
