--- 1_foundations/community_contributions/amirna2_contributions/personal-ai/career_chatbot_backup.py
@@ -1,7 +1,0 @@
-"""
-This is a backup of the original monolithic career_chatbot.py file.
-The original has been refactored into modules.
-"""
-
-# Content moved to career_chatbot_backup.py for reference
-# The original career_chatbot.py is now a compatibility wrapper
