--- 1_foundations/community_contributions/amirna2_contributions/personal-ai/chatbot/__init__.py
@@ -1,9 +1,0 @@
-"""
-Main chatbot module for the AI Career Assistant.
-"""
-
-from .main import CareerChatbot
-
-__all__ = [
-    'CareerChatbot'
-]
