--- 1_foundations/community_contributions/amirna2_contributions/personal-ai/core/__init__.py
@@ -1,11 +1,0 @@
-"""
-Core business logic for the AI Career Assistant.
-"""
-
-from .evaluator import Evaluator
-from .tools import ToolRegistry
-
-__all__ = [
-    'Evaluator',
-    'ToolRegistry'
-]
