--- 1_foundations/community_contributions/amirna2_contributions/personal-ai/models/__init__.py
@@ -1,15 +1,0 @@
-"""
-Data models for the AI Career Assistant.
-"""
-
-from .config import ChatbotConfig
-from .evaluation import Evaluation, StructuredResponse
-from .job_matching import SkillAssessment, JobMatchResult
-
-__all__ = [
-    'ChatbotConfig',
-    'Evaluation', 
-    'StructuredResponse',
-    'SkillAssessment',
-    'JobMatchResult'
-]
