--- 1_foundations/community_contributions/amirna2_contributions/personal-ai/models/evaluation.py
@@ -1,20 +1,0 @@
-"""
-Evaluation models for the AI Career Assistant.
-"""
-
-from typing import List
-from pydantic import BaseModel
-
-
-class Evaluation(BaseModel):
-    """Evaluation result for a response"""
-    is_acceptable: bool
-    feedback: str
-
-
-class StructuredResponse(BaseModel):
-    """Structured response with reasoning and evidence"""
-    response: str
-    reasoning: str
-    tools_used: List[str]
-    facts_used: List[str]
