--- 1_foundations/community_contributions/amirna2_contributions/personal-ai/services/__init__.py
@@ -1,13 +1,0 @@
-"""
-Services for the AI Career Assistant.
-"""
-
-from .notification import NotificationService
-from .web_search import WebSearchService
-from .document_loader import DocumentLoader
-
-__all__ = [
-    'NotificationService',
-    'WebSearchService', 
-    'DocumentLoader'
-]
